requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django==5.2.9
django-classy-tags==4.1.0
django-cms==5.0.5
django-filer==3.4.1
django-formtools==2.5.1
django-fsm-2==4.1.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-polymorphic==4.4.1
django-sekizai==4.1.0
django-treebeard==4.8.0
djangocms-admin-style==3.3.1
djangocms-attributes-field==4.1.2
djangocms-file==3.0.1
djangocms-googlemap==2.2.0
djangocms-link==5.1.0
djangocms-picture==4.1.1
djangocms-snippet==5.0.2
djangocms-style==3.1.0
djangocms-text-ckeditor==5.1.7
djangocms-versioning==2.5.0
djangocms-video==3.1.0
easy-thumbnails==2.10.1
html5lib==1.1
lxml==6.0.2
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
reportlab==4.4.6
six==1.17.0
sqlparse==0.5.4
svglib==1.5.1
tinycss2==1.5.1
typing_extensions==4.15.0
webencodings==0.5.1
